Splitting dogboneII implementation into its constituents
